app/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
openpyxl>=3.0.0
Pillow>=9.0.0
PyYAML>=6.0
docker>=6.0.0
kubernetes>=26.0.0
bcrypt>=4.0.0
google-cloud-firestore>=2.11.0